remove pay.reddit.com as a redirected url
This seems to have tripped some check google has as they are sending me
emails telling me I need to add a privacy policy as my extension deals
with sensitive user data such as payment details. Apparently the
extension will be taken down if this is not dealt with.

I'm hoping simply removing this line will be enough.