feat: drop Accept header from i.redd.it requests
When loading a page as a document, browsers put HTML and other MIME
types at the beginning of the Accept headers.
This is not the case for <img> tags.

Reddit abuses this to serve an HTML page on i.redd.it instead of the
actual image when navigating directly to its URL.
Dropping the header restores the previous, sane behavior.

Fixes: #78